feat: refactor utils
