+++ src/mosk/destination.js
@@ -1,5 +1,5 @@
 import { descriptions, names, MAX_DESTINATIONS_COUNT } from './const.js';
-import { getRandomInteger, getRandomArrayElement } from '../utils.js';
+import { getRandomInteger, getRandomArrayElement } from '../utils/common.js';
 
 
 const getPhotos = () => {
